cdk/requirements.txt
aws-cdk-lib
constructs>=10.0.0
pydantic>=2.0.0
boto3>=1.26.0